Fix QASM bug There were two bugs in using QASM with the program.
The first bug was that the temporary files couldn't find each other by
name, as temporary files have a bunch of garbage tacked on to the end of
their names upon creation. This was fixed by formatting the template to
use the file object's path, with correct escapes.

The second bug was that the measurement at the end was using incorrect
indices due to an overzealous "fixQASM" function to decrement all the
register indices by the computed offset. Combined with a bug in the
function numQubits to find the size of the circuit, there were negative
indices in the translated QASM. To fix this, I fixed numQubits and added
the offset to the generated measurement registers before calling the fix
function.